# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto - OpenAPI 3.0
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! \
    Directly from your PC, you can upload your photos, and they will be visible to everyone who is following you.
    
    Functional design specification:
    
    - The login is performed by specifying the username and it returns a crafted userID(token).
    - Once in, the user sees on the web app the stream of photos of the users that he follows ordered in reverse chronological order.
    - Each photo has information about the date-time it was uploaded, the number of likes and the number of comments.
    - Users can place a like to photos from other users(not from themselves);
    - Users can remove likes from photos of other users;
    - Users can add comments to any photo and can edit comments to any photo(even for themselves);
    - Only the author can remove comments from a photo;
    - Users have their own personal profiles.
    
    When you access to the personal profile of a user you can see:
    - user’s photos (in reverse chronological order), 
    - their followers,
    - their following, 
    - how many photos have been uploaded.
    
    User action:
    - change his/her own username,
    - follow/unfollow other users,
    - upload photos, 
    - remove photos: also remove likes and comments.
    - Users can ban other users.
...
    Alice can decide to remove the ban at any moment.
    
...
    
...
    
...
servers:
...
components:
  parameters:
    user_id:
      description: User whose profile belongs to. 
      name: user_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UserID" }
    photo_id:
      description: ID of the picture considered
      name: photo_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PhotoID" }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ShortProfile:
        description: Represents the information displayed in the list of likes, followers, ecc...
        type: object
        properties:
          username: {$ref: "#/components/schemas/Username"}
          profile_picture_url:
            description: URL of the profile picture. Accepting only ID attached with .png/.jpg/.jpeg extensions.
            type: string
            pattern: '^[a-zA-Z0-9-_]*\.(png|jpg|jpeg)$'
            example: "Cksvy-ZXtXRE.jpeg"
            minLength: 0
            maxLength: 150
    Profile:
      description: Represents the information seen in the Profile Page of a user
      type: object
      properties:
        userId: {$ref: "#/components/schemas/UserID"}
        username: {$ref: "#/components/schemas/Username"}
        pictures_count:
          description: Number of photos in the profile of the user
          type: integer
          example: 169
        followers_count:
          description: Number of users that follow the profile
          type: integer
          example: 2567
        follows_count:
          description: number of users that the user follows
          type: integer
          example: 689
        profile_picture_url:
          description: URL of the profile picture. Accepting only ID attached with.png/.jpg/.jpeg extensions.
          type: string
          pattern: '^[a-zA-Z0-9-_]*\.(png|jpg|jpeg)$'
          example: "Cksvy-ZXtXRE.jpg"
          minLength: 0
          maxLength: 150
        bio:
          description: Biography of the profile. |
            I'm keeping the pattern simple. Not doing crazy things. Just allowing alphanumeric characters and basic punctuation.
          type: string
          pattern: '^[a-zA-Z0-9,._:;?!\x27\- ]*$'
          example: "Discover what's next"
          # on WASAPhoto
          minLength: 0
          maxLength: 150
    Photo:
      description: Attributes of a photo
      type: object
      properties:
        photoId: {$ref: "#/components/schemas/PhotoID"}
        timestamp:
          description: Date and time of creation following RFC3339
          type: string
          format: date-time
          example: 2022-11-04T16:35:22-04:00
          minLength: 20
          maxLength: 25
        likes_count:
          description: Number of likes
          type: integer
          example: 169
        comments_count:
          description: Number of comments
          type: integer
          example: 15
        image:
          description: |
            URL of the image just uploaded.
            Accepting only ID attached with .png/.jpg/.jpeg extensions.
          type: string
          pattern: '^[a-zA-Z0-9-_]*\.(png|jpg|jpeg)$'
          example: "0a25d5f2-3b07-492e-ab87-406c4e7422b8.png"
          minLength: 0
          maxLength: 150
        caption:
          description: A written description or explanation about a photo to provide more context
          type: string
          pattern: '^[a-zA-Z0-9,._:;?!\x27\- ]*$'
          example: "Vodka may not be the answer but it's worth a shot."
          minLength: 0
          maxLength: 100
        username: {$ref: "#/components/schemas/Username"}
...
    UserID:
      description: |
        The simplified login tell us that, since there is no password, the username must be unique among the application. 
        Logging in with an existing name will grant access to that user information. 
        Hence it would make sense to use the username as ID since we don't have to represent real people but virtual people, namely usernames, by simplifying in this way the ReST OpenAPI. 
        However, since it's not applicable in a real life scenario, I will use a userID that's different from the username. Thus, I'm choosing 20 chars long strings because they can store several billions of information(64^20, since the alphabet is "[a-z][A-Z][0-9]-_", 26+26+10+2=64), they are nicer to read then integers since they are shorter and almost equally fast to compute.
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 1
      maxLength: 20
      example: "U3dhZ2dlciByb2Nrcw"
      readOnly: true
    PhotoID:
      description: Unique photo identifier
      type: string
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 1
      maxLength: 20
      example: "CksvyZXtXRE"
      readOnly: true
...
    Username:
      description: It represents the name of the user in WASA Photo
      type: string
      example: LeoMessi88
      pattern: '^[a-zA-Z0-9-_]*$'
      minLength: 3
      maxLength: 16
  responses:
...
    InternalServerError: 
      description: |-
        The server encountered an internal error.
        Further info in server logs.
    UnauthorizedError:
      description: Access token is missing or invalid.
paths:
...
  /users/:
    get:
      tags:
        - User
      summary: Find user by username
      description: |
        The profile of an user contains the username, the number of photos it has, the number of follows/followers, a bio and a profile picture.
      operationId: getUserProfile
      parameters:
        - name: username
          in: query
          description: Username to filter by
          required: false
          explode: false
          schema:
              $ref: '#/components/schemas/Username'
      responses:
        '200':
          description: Successful operation. User profile given in the response body.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Profile'
        "400": { description: Invalid username supplied }
        "404": { description: User not found }
        "500": { $ref: "#/components/responses/InternalServerError" }
      security:
        - bearerAuth: []
...
  /users/{user_id}/followers/{follower_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - name: follower_id
        in: path
        required: true
        description: User who is following
        schema:
          { $ref: "#/components/schemas/UserID" }

    put:
      tags:
        - User
      summary: Follow user
      description: >
        Following an user imply incrementing his followers count by one, our follows count by one and to add his profile in our stream. I'm using PUT for its idempotency.
      operationId: followUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserID'
      responses:
        "201":
          description: Follow created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserID'
        "204":
          description: Already following current user.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "409": { description: An user can't follow himself. }
      security:
        - bearerAuth: []

    delete:
      tags:
        - User
      summary: Unfollow user
      description: This can only be done by the logged in user.
      operationId: unfollowUser
      responses:
        "204": { description: User unfollowed successfully. }
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "409": { description: An user can't unfollow himself. }
      security:
        - bearerAuth: []
  /users/{user_id}/stream/:
    parameters:
      - $ref: "#/components/parameters/user_id"

    get:
      tags:
      - Stream
      summary: Stream user
      description: >
        Homepage of the WASAPhoto web app that corresponds to a stream of photo of the users the authenticated user is following
      operationId: getMyStream
      responses:
        "200":
          description: Array of photos ordered in reverse chronological order of the users we are following
          content:
            application/json: 
              schema:
                description: |
                  Represents the stream of photos in the user's homepage.
                  Users can see comments in batches of 20.
                type: array
                items: 
                  $ref: "#/components/schemas/Photo"
                minItems: 0
                maxItems: 20
                uniqueItems: true
        "204": { description: No content. User still doesn't follow any user.}
      security:
        - bearerAuth: []
...
  /users/{user_id}/bans/:
    parameters:
      - $ref: "#/components/parameters/user_id"
    
    get:
      tags:
        - User
      summary: See banned users
      description: |
         List of users that has been banned from this profile
      operationId: getBannedUsers
      responses:
        "200":
          description: |
            List of banned users with their respective profile photo. At most 5% of the user's followers can be banned.
          content:
            application/json:
              schema:
                description: |
                  Represents the users that are banned from the profile.
                  Users can see banned users in batches of 50.
                type: array
                minItems: 0
                maxItems: 50
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/ShortProfile'
      security:
          - bearerAuth: []
...
  /photos/{photo_id}:
    description: >
      This resource represents an individual photo in the system belonging to the user user-ID.
      Each photo is identified by a numeric `ID`.
      
    parameters:
      - $ref: "#/components/parameters/photo_id"
      
    get:
      tags:
       - Profile
      summary: Get a photo
      description: Retrieve a single photo
      operationId: getUserPhoto
      responses: 
        "200":
          description: Photo found
          content:
            application/json: 
              schema:
                { $ref: "#/components/schemas/Photo" }
        "204": { description: No content}
        "404": { description: Not found }
      security:
        - bearerAuth: []
      
    delete:
      tags:
          - Profile
      summary: Deletes a photo
      description: Deletes a photo from a user profile. It also removes like and comments.
      operationId: deletePhoto
      responses:
        "204": { description: Picture deleted successfully }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "404": { description: Picture not found }
        "500": { $ref: "#/components/responses/InternalServerError" }
      security:
          - bearerAuth: []
...
  /photos/{photo_id}/likes/{like_id}:
    parameters:
      - $ref: "#/components/parameters/photo_id"
      - name: like_id
        in: path
        required: true
        description: User who is liking the photo
        schema:
          { $ref: "#/components/schemas/UserID" }
    
    put:
      tags:
        - Likes
      summary: Like a photo
      description: Add a like on a photo. I'm using PUT for its idempotency.
      operationId: likePhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserID'
      responses:
        "201":
          description: Like created.
          content:
            application/json: 
              schema:
                { $ref: "#/components/schemas/UserID" }
        "204":
          description: No content. Like already inserted.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "409": { description: An user can't like its own photos. }
      security:
        - bearerAuth: []


    delete:
      tags:
        - Likes
      summary: Unlike a photo
      description: Remove a like from a photo
      operationId: unlikePhoto
      responses:
        "204":
          description: No content. Like removed.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "409": { description: An user can't unlike its own photos. }
      security:
        - bearerAuth: []
...